Create reusable caching traits.
